packages/forecaster/pyproject.toml
[project]
name = "forecaster"
version = "0.2.0"
description = "Forecast Studio — core forecasting and agent library"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "statsmodels>=0.14.0",
    "scikit-learn>=1.3.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "mlforecast>=0.13.0",
    "lightgbm>=4.1.0",
    "prophet>=1.1.5",
    "langgraph>=0.2.0",
    "langfuse>=4.0.0",
]

[project.optional-dependencies]
duckdb = ["duckdb>=0.9.0"]

[project.scripts]
forecast-studio = "forecaster.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["forecaster"]

[tool.mypy]
[[tool.mypy.overrides]]
module = ["openai.*", "statsmodels.*"]
ignore_missing_imports = true
